fix: fix merge conflicts after rebase
