+++ apps/users/forms/__init__.py
@@ -1,0 +1,1 @@
+from .users_form import CustomUserCreationForm
