+++ apps/users/forms/users_form.py
@@ -1,0 +1,9 @@
+from django import forms
+from django.contrib.auth.forms import UserCreationForm
+from ..models import User
+
+
+class CustomUserCreationForm(UserCreationForm):
+    class Meta:
+        model = User
+        fields = ("username", "password1", "password2")  # 添加你的自訂字段
