+++ apps/users/templates/users/home.html
@@ -1,6 +1,17 @@
 {% extends "layouts/base.html" %} {% block content %}
 <h1>首頁</h1>
 
+{% if request.user.is_authenticated %}
+<h2>{{ request.user }}</h2>
+<form action="{% url 'users:sign_out' %}" method="POST">
+  {% csrf_token %} 
+  <button>登出</button>
+</form>
+
+{% endif %}
+
+{% if not request.user.is_authenticated %}
 <a class="link" href="{% url 'users:sign_in' %}">登入</a>
 <a class="link" href="{% url 'users:register' %}">註冊</a>
+{% endif %}
 {% endblock %}
