+++ apps/users/templates/users/sign_in.html
@@ -28,6 +28,7 @@
       >領好禮吧！
     </div>
   </form>
+  <a href="{% url 'social:begin' 'line' %}">使用 LINE 登入</a>
 </section>
 
 {% endblock %}
