--- page/templates/index.html
@@ -1,36 +1,0 @@
-{% extends 'layout/base.html' %}
-
-{% block content %}
-<h1>首頁</h1>
-
-{% if request.user.is_authenticated %}
-<h2>{{ request.user }}</h2>
-<form action="{% url 'page:sign_out' %}" method="POST">
-  {% csrf_token %} 
-  <button>登出</button>
-</form>
-
-{% endif %}
-
-
-{% if not request.user.is_authenticated %}
-<form action="{% url 'page:sign_in' %}" method="POST" autocomplete="off">
-  {% csrf_token %} 
-
-  <div class="form-control">
-    <label class="label">帳號</label>
-    <input class="input" type="text" name="username" />
-  </div>
-
-  <div class="form-control">
-    <label class="label">密碼</label>
-    <input class="input" type="password" name="password" />
-  </div>
-  
-  <div>
-    <button class="btn btn-primary">登入</button>
-  </div>
-</form>
-<a href="{% url 'social:begin' 'line' %}">使用 LINE 登入</a>
-{% endif %}
-{% endblock%}
